bug(UIKIT-982): Устранение циклических зависимостей
