+++ packages/components/src/Badge/types.ts
@@ -1,3 +1,7 @@
+import { BadgeProps as MuiBadgeProps } from '@mui/material/Badge/Badge';
+
+import { WithoutEmotionSpecific } from '../types';
+
 export type BadgeColor =
   | 'primary'
   | 'grey'
@@ -6,3 +10,17 @@ export type BadgeColor =
   | 'errorLight'
   | 'error'
   | 'success';
+
+export type BadgeProps = Omit<
+  WithoutEmotionSpecific<MuiBadgeProps>,
+  'color'
+> & {
+  /**
+   * Цвет фона
+   */
+  color: BadgeColor;
+  /**
+   * Наличие белой рамки
+   */
+  withBorder?: boolean;
+};
