+++ packages/components/src/CheckableTag/index.ts
@@ -1,1 +1,3 @@
 export * from './CheckableTag';
+
+export * from './types';
