+++ packages/components/src/Checkbox/Checkbox.tsx
@@ -1,20 +1,8 @@
 import { forwardRef } from 'react';
-import { CheckboxProps as MuiCheckboxProps } from '@mui/material/Checkbox/Checkbox';
-
-import { WithoutEmotionSpecific } from '../types';
 
 import { CheckboxStyled } from './styles';
 import { CheckedIcon, DefaultIcon, IndeterminateIcon } from './icons';
-
-export type CheckboxProps = Omit<
-  WithoutEmotionSpecific<MuiCheckboxProps>,
-  'size' | 'color'
-> & {
-  /**
-   * Флаг для активации error стилей.
-   */
-  isError?: boolean;
-};
+import { CheckboxProps } from './types';
 
 export const Checkbox = forwardRef<HTMLButtonElement, CheckboxProps>(
   (
