+++ packages/components/src/Checkbox/index.ts
@@ -1,1 +1,3 @@
 export * from './Checkbox';
+
+export * from './types';
