+++ packages/components/src/Chevron/index.ts
@@ -1,1 +1,3 @@
 export * from './Chevron';
+
+export * from './types';
