+++ packages/components/src/Grid/GridComponent.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from 'react';
 
-import { GridProps } from './Grid';
+import { GridProps } from './types';
 
 export const GridComponent = forwardRef<HTMLDivElement, GridProps>(
   ({ component: Component = 'div', ...props }, ref) => {
