+++ packages/components/src/OverflowTypography/index.ts
@@ -1,1 +1,3 @@
 export * from './OverflowTypography';
+
+export * from './types';
