+++ packages/components/src/Radio/index.ts
@@ -1,1 +1,3 @@
 export * from './Radio';
+
+export * from './types';
