+++ packages/components/src/Tag/types.ts
@@ -1,6 +1,8 @@
 import { ReactElement } from 'react';
+import { ChipProps as MuiTagProps } from '@mui/material/Chip/Chip';
 
 import { BadgeColor } from '../Badge';
+import { WithoutEmotionSpecific } from '../types';
 
 import { TagColors, TagStates, TagVariants } from './enums';
 
@@ -17,3 +19,35 @@ export type TagAddonProps = {
 };
 
 export type TagAddon = (props: TagAddonProps) => ReactElement | null;
+
+export type TagProps = Omit<
+  WithoutEmotionSpecific<MuiTagProps>,
+  'color' | 'variant' | 'size'
+> & {
+  /**
+   * Цвет тега
+   */
+  color?: TagColor;
+  /**
+   * Тип тега
+   */
+  variant?: TagVariant;
+  /**
+   * Скругленная форма тега
+   */
+  rounded?: boolean;
+  /**
+   * Размер тега
+   */
+  size?: TagSize;
+
+  /**
+   * Контент слева от label
+   */
+  startAddon?: TagAddon;
+
+  /**
+   * Контент справа от label
+   */
+  endAddon?: TagAddon;
+};
